only wait for 32 blocks with empty check
