+++ ethereum/contracts.go
@@ -6,13 +6,9 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/mezo-org/mezod/ethereum/bindings/portal"
-	"github.com/mezo-org/mezod/ethereum/bindings/portal/sepolia/gen/abi"
 )
 
 type BridgeContract interface {
-	ValidateAssetsUnlocked(
-		assetsUnlocked abi.MezoBridgeAssetsUnlocked,
-	) (bool, error)
 	AttestBridgeOut(
 		assetsUnlocked *portal.MezoBridgeAssetsUnlocked,
 	) (*types.Transaction, error)
